Move functionality to CiscoTelnetSession class
